Updating appveyor scripts

+++ deploy.bat
@@ -13,7 +13,7 @@ rem Deploy DLLs
 %LIB_DIR%\windeployqt.exe --force --release --dir %OUTPUT_DIR% --no-quick-import --no-translations --no-webkit2 --no-angle --no-svg --no-webkit %EXE_PATH%
 
 rem Removing unused DLLs
-del /q %OUTPUT_DIR%\opengl32sw.dll
+if exist %OUTPUT_DIR%\opengl32sw.dll del /q %OUTPUT_DIR%\opengl32sw.dll
 
 rem Deploy Translations
 for %%l in (%LANGUAGES%) do (
@@ -25,6 +25,3 @@ xcopy /y %LIB_DIR%\liblz4.dll %OUTPUT_DIR%
 
 rem Deploy Exe
 xcopy /y %EXE_PATH% %OUTPUT_DIR%
-
-rem Compress Exe and DLLs. Note: DLLs in platforms/ directory should not be compressed.
-rem upx %OUTPUT_DIR%\*.dll %OUTPUT_DIR%\Deling.exe
